requirements.txt
websockets>=10
aiosqlite>=0.21
fastapi>=0.110
uvicorn[standard]>=0.23
staticmap>=0.5
Pillow>=10
pytest>=8
httpx>=0.27
